requirements.txt
# Dependencias para nITROGEN Linux

# MQTT client para comunicación con brokers MQTT
paho-mqtt>=1.6.1

# Si en el futuro necesitas conexiones HTTP/REST
requests>=2.31.0

# Para trabajar con archivos JSON de configuración
# (ya viene con Python, pero bueno tenerlo explícito si usas jsonschema)
jsonschema>=4.20.0

# Para formatear y validar fechas avanzadas (complemento a datetime)
python-dateutil>=2.8.2

# Para interfaces de línea de comandos más avanzadas (opcional)
click>=8.1.7

# Para logging más robusto (opcional pero recomendado)
colorlog>=6.8.0
